Rename python project to zimtohrli
